requirements.txt
psycopg2-binary==2.9.9
Pillow>=10.2.0
bcrypt==4.0.1
PyQt5==5.15.10